3785 - getTableData: extract common code
