+++ src/server/controller/cycleData/getTableData.ts
@@ -1,49 +1,14 @@
 import { Objects } from 'utils/objects'
 
-import { Areas, Country, CountryIso } from 'meta/area'
-import { Assessment, Cycle, TableName, TableNames, VariableCache } from 'meta/assessment'
-import { RecordAssessmentData, RecordCountryData, TablesCondition } from 'meta/data'
+import { Country, CountryIso } from 'meta/area'
+import { TableName, TableNames } from 'meta/assessment'
+import { RecordAssessmentData, RecordCountryData } from 'meta/data'
 
+import { getTablesCondition } from 'server/controller/cycleData/tableData/getTablesCondition'
+import { Props } from 'server/controller/cycleData/tableData/props'
 import { BaseProtocol, DB } from 'server/db'
 import { CountryRepository } from 'server/repository/assessmentCycle/country'
-import { DataRepository } from 'server/repository/assessmentCycle/data'
 import { DataRedisRepository } from 'server/repository/redis/data'
-
-type Props = {
-  assessment: Assessment
-  cycle: Cycle
-  countryISOs: Array<CountryIso>
-  tableNames: Array<string> // TODO: refactor use TablesCondition instead
-  variables?: Array<string>
-  columns?: Array<string>
-  mergeOdp?: boolean
-  /**
-   * @deprecated
-   * Merge dependencies to tables condition
-   * TODO: Handle dependencies differently (currently only used in calculateNode -> part of updateDependencies job)
-   */
-  dependencies?: Array<VariableCache>
-}
-
-const _getTablesCondition = (
-  props: Pick<Props, 'tableNames' | 'columns' | 'variables' | 'mergeOdp'>
-): TablesCondition => {
-  const { tableNames, columns, variables, mergeOdp } = props
-
-  const tables: TablesCondition = {}
-
-  tableNames.forEach((tableName) => {
-    tables[tableName] = { columns, variables }
-  })
-  const withOdp =
-    mergeOdp &&
-    (tableNames.includes(TableNames.extentOfForest) || tableNames.includes(TableNames.forestCharacteristics))
-  if (withOdp) {
-    tables[TableNames.originalDataPointValue] = { columns, variables }
-  }
-
-  return tables
-}
 
 const _mergeODPTable = (props: {
   countryIso: CountryIso
@@ -60,24 +25,9 @@ const _mergeODPTable = (props: {
 }
 
 export const getTableData = async (props: Props, client: BaseProtocol = DB): Promise<RecordAssessmentData> => {
-  const { assessment, columns, countryISOs, cycle, mergeOdp: mergeOdpProp, tableNames, variables } = props
+  const { assessment, columns, countryISOs, cycle, mergeOdp, tableNames, variables } = props
 
-  const isRegion = !Areas.isISOCountry(countryISOs[0])
-  // Ignore mergeOdp if request is for region
-  const mergeOdp = mergeOdpProp && !isRegion
-
-  const tables = _getTablesCondition({ tableNames, columns, variables, mergeOdp })
-
-  const aggregate = isRegion
-
-  // TODO: Cache aggregated Table data
-  if (aggregate) {
-    return {
-      [assessment.props.name]: {
-        [cycle.name]: await DataRepository.getAggregatedTableData({ assessment, cycle, countryISOs, tables }, client),
-      },
-    }
-  }
+  const tables = getTablesCondition({ tableNames, columns, variables, mergeOdp })
 
   const tableData = await DataRedisRepository.getCountriesData({ assessment, cycle, tables, countryISOs })
