+++ src/server/controller/cycleData/tableData/index.ts
@@ -1,5 +1,7 @@
+import { getAggregatedTableData } from './getAggregatedTableData'
 import { refreshViews } from './refreshViews'
 
 export const TableData = {
+  getAggregatedTableData,
   refreshViews,
 }
